+++ test/runtests.jl
@@ -425,7 +425,6 @@ end
         @test (BoundingBox(1,2,3,4) ⊆ BoundingBox(1,2,3,4)) == true
         @test (BoundingBox(1,2,3,4) ⊆ BoundingBox(1,4,0,3)) == false
         @test (BoundingBox(0,2,3,4) ⊆ BoundingBox(1,2,3,5)) == false
-        @test_deprecated (BoundingBox(0,2,3,4) ∈ BoundingBox(1,2,3,5)) == false
 
         C = Point(x = 0.5*(B.xmin + B.xmax), y = 0.5*(B.ymin + B.ymax))
         @test center(B) === C